Remove original CreateNewRec.tsx (not in use) and rename NewModal.tsx to CreateNewRec.tsx
